requirements.txt
aiogram>=3.0
aiosqlite
APScheduler
python-dotenv
